+++ extensions/wrapper/wrapper-common-api/src/main/java/de/sovity/edc/ext/wrapper/api/common/model/UiPolicyCreateRequest.java
@@ -15,7 +15,6 @@
 package de.sovity.edc.ext.wrapper.api.common.model;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.media.Schema.RequiredMode;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -23,26 +22,17 @@ import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 
-/**
- * Opinionated subset of the EDC policy.
- *
- * @author tim.dahlmanns@isst.fraunhofer.de
- */
+import java.util.List;
+
 @Getter
 @Setter
 @ToString
 @AllArgsConstructor
 @Builder(toBuilder = true)
 @RequiredArgsConstructor
-@Schema(description = "Type-Safe OpenAPI generator friendly Policy DTO that supports an opinionated"
+@Schema(description = "Type-Safe OpenAPI generator friendly Policy Create DTO that supports an opinionated"
         + " subset of the original EDC Policy Entity.")
-public class PolicyDto {
-
-    @Schema(description = "Legacy JSON as built by the Management API. Will be replaced "
-            + "in the future by a type-safe variant without polymorphisms that can be used "
-            + "for our generated clients.", requiredMode = Schema.RequiredMode.REQUIRED)
-    private String legacyPolicy;
-
-    @Schema(description = "Permission for this policy", requiredMode = RequiredMode.REQUIRED)
-    private PermissionDto permission;
+public class UiPolicyCreateRequest {
+    @Schema(description = "Conjunction of required expressions for the policy to evaluate to TRUE.")
+    private List<UiPolicyConstraint> constraints;
 }