+++ extensions/wrapper/wrapper/src/main/java/de/sovity/edc/ext/wrapper/WrapperExtensionContextBuilder.java
@@ -15,6 +15,12 @@
 package de.sovity.edc.ext.wrapper;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import de.sovity.edc.ext.wrapper.api.common.mappers.OperatorMapper;
+import de.sovity.edc.ext.wrapper.api.common.mappers.PolicyMapper;
+import de.sovity.edc.ext.wrapper.api.common.mappers.utils.AtomicConstraintMapper;
+import de.sovity.edc.ext.wrapper.api.common.mappers.utils.ConstraintExtractor;
+import de.sovity.edc.ext.wrapper.api.common.mappers.utils.LiteralMapper;
+import de.sovity.edc.ext.wrapper.api.common.mappers.utils.PolicyValidator;
 import de.sovity.edc.ext.wrapper.api.ee.EnterpriseEditionResourceImpl;
 import de.sovity.edc.ext.wrapper.api.ui.UiResource;
 import de.sovity.edc.ext.wrapper.api.ui.pages.asset.AssetApiService;
@@ -22,12 +28,16 @@ import de.sovity.edc.ext.wrapper.api.ui.pages.asset.services.AssetBuilder;
 import de.sovity.edc.ext.wrapper.api.ui.pages.asset.services.utils.AssetPropertyMapper;
 import de.sovity.edc.ext.wrapper.api.ui.pages.contracts.ContractAgreementPageApiService;
 import de.sovity.edc.ext.wrapper.api.ui.pages.contracts.ContractAgreementTransferApiService;
+import de.sovity.edc.ext.wrapper.api.ui.pages.contracts.ContractDefinitionApiService;
 import de.sovity.edc.ext.wrapper.api.ui.pages.contracts.services.ContractAgreementDataFetcher;
 import de.sovity.edc.ext.wrapper.api.ui.pages.contracts.services.ContractAgreementPageCardBuilder;
+import de.sovity.edc.ext.wrapper.api.ui.pages.contracts.services.ContractDefinitionBuilder;
 import de.sovity.edc.ext.wrapper.api.ui.pages.contracts.services.TransferProcessStateService;
 import de.sovity.edc.ext.wrapper.api.ui.pages.contracts.services.TransferRequestBuilder;
 import de.sovity.edc.ext.wrapper.api.ui.pages.contracts.services.utils.ContractAgreementUtils;
 import de.sovity.edc.ext.wrapper.api.ui.pages.contracts.services.utils.ContractNegotiationUtils;
+import de.sovity.edc.ext.wrapper.api.ui.pages.contracts.services.utils.CriterionMapper;
+import de.sovity.edc.ext.wrapper.api.ui.pages.policy.PolicyDefinitionApiService;
 import de.sovity.edc.ext.wrapper.api.ui.pages.transferhistory.TransferHistoryPageApiService;
 import de.sovity.edc.ext.wrapper.api.ui.pages.transferhistory.TransferHistoryPageAssetFetcherService;
 import de.sovity.edc.ext.wrapper.api.usecase.UseCaseResource;
@@ -40,7 +50,9 @@ import org.eclipse.edc.connector.contract.spi.offer.store.ContractDefinitionStor
 import org.eclipse.edc.connector.policy.spi.store.PolicyDefinitionStore;
 import org.eclipse.edc.connector.spi.asset.AssetService;
 import org.eclipse.edc.connector.spi.contractagreement.ContractAgreementService;
+import org.eclipse.edc.connector.spi.contractdefinition.ContractDefinitionService;
 import org.eclipse.edc.connector.spi.contractnegotiation.ContractNegotiationService;
+import org.eclipse.edc.connector.spi.policydefinition.PolicyDefinitionService;
 import org.eclipse.edc.connector.spi.transferprocess.TransferProcessService;
 import org.eclipse.edc.connector.transfer.spi.store.TransferProcessStore;
 import org.eclipse.edc.policy.engine.spi.PolicyEngine;
@@ -75,6 +87,8 @@ public class WrapperExtensionContextBuilder {
             PolicyDefinitionStore policyDefinitionStore,
             PolicyEngine policyEngine,
             TransferProcessStore transferProcessStore,
+            ContractDefinitionService contractDefinitionService,
+            PolicyDefinitionService policyDefinitionService,
             TransferProcessService transferProcessService,
             TypeTransformerRegistry transformerRegistry,
             ContractNegotiationObservable negotiationObservable
@@ -94,6 +108,10 @@ public class WrapperExtensionContextBuilder {
                 contractAgreementDataFetcher,
                 contractAgreementPageCardBuilder
         );
+        var operatorMapper = new OperatorMapper();
+        var criterionMapper = new CriterionMapper(operatorMapper);
+        var contactDefinitionBuilder = new ContractDefinitionBuilder(criterionMapper);
+        var contractDefinitionApiService = new ContractDefinitionApiService(contractDefinitionService, criterionMapper, contactDefinitionBuilder);
         var transferHistoryPageApiService = new TransferHistoryPageApiService(
                 assetService,
                 contractAgreementService,
@@ -118,12 +136,21 @@ public class WrapperExtensionContextBuilder {
                 transferRequestBuilder,
                 transferProcessService
         );
+        var jsonLdObjectMapper = new ObjectMapper();
+        var literalMapper = new LiteralMapper(objectMapper);
+        var atomicConstraintMapper = new AtomicConstraintMapper(literalMapper, operatorMapper);
+        var policyValidator = new PolicyValidator();
+        var constraintExtractor = new ConstraintExtractor(policyValidator, atomicConstraintMapper);
+        var policyMapper = new PolicyMapper(jsonLdObjectMapper, constraintExtractor, atomicConstraintMapper);
+        var policyDefinitionApiService = new PolicyDefinitionApiService(policyDefinitionService, policyMapper);
         var uiResource = new UiResource(
                 contractAgreementApiService,
                 contractAgreementTransferApiService,
                 transferHistoryPageApiService,
                 transferHistoryPageAssetFetcherService,
-                assetApiService
+                assetApiService,
+                policyDefinitionApiService,
+                contractDefinitionApiService
         );
 
         // Use Case API